LPS-87755 Rename to sections since we are also going to be adding elements
