+++ modules/apps/layout/layout-admin-web/src/main/resources/META-INF/resources/js/fragments_editor/components/sidebar/FragmentsEditorSidebarContent.soy
@@ -26,8 +26,8 @@
 
 	<div class="fragments-editor-sidebar-content">
 		<div class="fragments-editor-sidebar-section">
-			{if $_sectionId == 'fragments'}
-				{call com.liferay.layout.admin.web.SidebarFragmentsSection.render}
+			{if $_sectionId == 'sections'}
+				{call com.liferay.layout.admin.web.SidebarSectionsSection.render}
 					{param fragmentEntryLinks: $fragmentEntryLinks /}
 					{param layoutData: $layoutData /}
 					{param sections: $sections /}
