+++ modules/apps/layout/layout-admin-web/src/main/resources/META-INF/resources/js/fragments_editor/components/sidebar/fragments/SidebarAvailableSections.soy
@@ -1,7 +1,7 @@
-{namespace com.liferay.layout.admin.web.SidebarAvailableFragments}
+{namespace com.liferay.layout.admin.web.SidebarAvailableSections}
 
 /**
- * SidebarAvailableFragments
+ * SidebarAvailableSections
  */
 {template .render}
 	{@param sections: list<[
